requirements.txt
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2==2.9.7
